eliminate spurious newlines
The spurious newlines were caused by needing to _ensure_ that the demarcaters were on separate lines.
We could insert a newline at the end of the demarcater lines,
but we would also have to insert one at the start of the "end" demarcater
since it isn't guaranteed that every repl line ends with a newline.
But when a repl line doesn't print anything, there are now 2 newlines in a row,
resulting in an apparenty empty line.

This approach does some basic parsing of the line.
It fully encloses the demarcaters in pilcrows,
which allows us to easily distinguish them.
The parser is honestly pretty dumb, but it doesn't need to be better.

We also allow the command character to be customisable, because why not